Add script to clean repo data with GitHub API
This adds a script that queries the GitHub repo API to prune known-github-tomls.jsonl into known-github-tomls-clean.jsonl.

This solves three problems that i ran into when trying to scale the ecosystem CI with known-github-tomls.jsonl:

The original known-github-tomls.jsonl contains forks, which inflate some data incorrectly. There, we check in the API response whether the repo is a fork and exclude those.

Sometimes the `ref` is wrong, which would make my clones fail. We check the GitHub API for the default branch and use those.

Some entries in known-github-tomls.jsonl do not contain a `path`. Those are filtered out.

To avoid running into GitHub API rate limits, the results are cached into repo_api_data.jsonl